requirements.txt
Django==1.10.6
djangorestframework==3.6.2
psycopg2==2.7.1
django-rest-swagger==2.1.2